Procedimiento buscarIdPaciente
Procedimiento para buscar id de un paciente en el array de pacientes

+++ src/Paciente.java
@@ -41,7 +41,7 @@ public class Paciente extends Usuario {
         this.peso = 0;
     }
 
-    public Paciente(String tipoDocumento, int numeroDocumento,
+    public Paciente(String tipoDocumento, long numeroDocumento,
                     String nombrePaciente, String apellidoPaciente,
                     long numeroTelefonicoPaciente, String correoPaciente,
                     String direccionDeDomicilio, String fechaDeNacimiento,
